refactor: rename folder 'pages'
