+++ src/app/home.tsx
@@ -1,14 +1,15 @@
 import { BottomTabMenu } from "@/components/BottomTabMenu";
-import { HomePage } from "@/pages/app";
+import { HomePage } from "@/pages";
 import { View } from "react-native";
 
 export default function HomeScreen() {
   return (
-    <View className="flex-1">
-      <View className="flex-1">
+    <View className='flex-1'>
+      <View className='flex-1'>
         <HomePage />
       </View>
       <BottomTabMenu />
     </View>
   );
 }
+
