--- src/pages/app/index.ts
@@ -1,3 +1,0 @@
-export * from "./HomePage";
-export * from "./LoginPage";
-export * from "./TransactionsPage";
